refactor(game-service): change folder structure, file naming and remove unnecessary code
