+++ services/game_service/Dockerfile
@@ -2,11 +2,7 @@ FROM node:20 AS builder
 WORKDIR /app
 COPY ./package*.json ./
 RUN npm install
-COPY ./modules/ ./modules/
-COPY ./plugins/ ./plugins/
-COPY ./routes/ ./routes/
-COPY ./types/ ./types/
-COPY ./app.ts ./
+COPY src/ ./src/
 COPY ./tsconfig.json ./
 RUN npx tsc
