+++ services/game_service/package.json
@@ -8,7 +8,7 @@
     "build": "tsc -p tsconfig.json",
     "build_watch": "tsc -w -p tsconfig.json",
     "dev": "node .",
-    "dev_watch": "tsx watch app.ts",
+    "dev_watch": "tsx watch src/app.ts",
     "start": "node ./dist/app.js"
   },
   "dependencies": {
