+++ services/game_service/src/models/ball.ts
@@ -1,6 +1,6 @@
-import { BallState, Point } from '../types/game.js';
-
-import { BALL_DIAMETER, BALL_INIT_SPEED, BALL_MAX_SPEED, PADDLE_HEIGHT, PADDLE_WIDTH } from '../types/constants.js';
+import { BALL_DIAMETER, BALL_INIT_SPEED, BALL_MAX_SPEED, PADDLE_HEIGHT, PADDLE_WIDTH } from '../types/game-constants.js';
+import {BallState} from "../types/ball.js";
+import {Point} from "../types/point.js";
 
 
 export class Ball {
