+++ services/game_service/src/routes/schemas/get-response.ts
