+++ services/game_service/src/services/rabbitmq-client.ts
