+++ services/game_service/src/types/ball.ts
@@ -1,0 +1,5 @@
+export interface BallState {
+    x: number;
+    y: number;
+    semidiameter: number;
+}
