+++ services/game_service/src/types/game-constants.ts
