+++ services/game_service/src/types/point.ts
@@ -1,0 +1,10 @@
+import {PaddleSide} from "./paddle.js";
+
+export interface CollisionPoint extends Point {
+    paddleSide: PaddleSide | null
+}
+
+export interface Point {
+    x: number;
+    y: number;
+}
