+++ services/game_service/src/types/websocket.ts
@@ -1,0 +1,15 @@
+import {WebSocket} from "@fastify/websocket";
+
+export interface WsParams {
+    gameId: string;
+}
+
+export interface WsQuery {
+    playerJWT: string;
+}
+
+export interface GameWebSocket extends WebSocket {
+    gameId: string;
+    playerSessionId: string;
+    username: string;
+}
