--- services/game_service/tests/index2.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Line Intersection Test</title>
-</head>
-<body style="background: black;">
-    <script type="module" src="../dist/modules/test.js"></script>
-</body>
-</html>
