--- services/game_service/tests/index_old.html
@@ -1,182 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Pong Game</title>
-    <style>
-        body {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            background: #1a1a1a;
-            color: white;
-            font-family: Arial, sans-serif;
-        }
-        canvas {
-            border: 1px solid white;
-            margin: 0px;
-        }
-        #controls {
-            margin: 20px;
-            padding: 10px;
-            border: 1px solid #333;
-            border-radius: 5px;
-        }
-        button {
-            margin: 5px;
-            padding: 8px 16px;
-            background: #333;
-            color: white;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-        }
-        button:hover {
-            background: #444;
-        }
-    </style>
-</head>
-<body>
-    <div id="controls">
-        <button onclick="connectPlayer1()">Connect as Player 1</button>
-        <button onclick="connectPlayer2()">Connect as Player 2</button>
-    </div>
-    <canvas id="gameCanvas" width="800" height="800"></canvas>
-
-    <script>
-        const canvas = document.getElementById('gameCanvas');
-        const ctx = canvas.getContext('2d');
-        let socket = null;
-        let gameState = null;
-        let isPlayer1 = false;
-
-        // Constants matching your backend
-        const PADDLE_HEIGHT = 15;
-        const PADDLE_WIDTH = 40;
-        const BALL_SEMIDIAMETER = 1;
-
-        function connectPlayer1() {
-            if (socket) socket.close();
-            socket = new WebSocket('ws://localhost:3000/api/games/ws?game_id=0b879657-b318-4159-b663-882d97f689dd&player_id=test1');
-            setupWebSocket();
-            isPlayer1 = true;
-        }
-
-        function connectPlayer2() {
-            if (socket) socket.close();
-            socket = new WebSocket('ws://localhost:3000/api/games/ws?game_id=0b879657-b318-4159-b663-882d97f689dd&player_id=test2');
-            setupWebSocket();
-            isPlayer1 = false;
-        }
-
-        function setupWebSocket() {
-            socket.onmessage = (event) => {
-                gameState = JSON.parse(event.data);
-                drawGame();
-            };
-
-            socket.onerror = (error) => {
-                console.error('WebSocket error:', error);
-            };
-
-            socket.onclose = () => {
-                console.log('WebSocket connection closed');
-            };
-        }
-
-        // Convert game coordinates (0-100) to canvas coordinates
-        function scaleX(x) {
-            return (x / 100) * canvas.width;
-        }
-
-        function scaleY(y) {
-            return (y / 100) * canvas.height;
-        }
-
-        function drawGame() {
-            if (!gameState) return;
-
-            // Clear canvas
-            ctx.fillStyle = '#1a1a1a';
-            ctx.fillRect(0, 0, canvas.width, canvas.height);
-
-            // Draw middle line
-            ctx.strokeStyle = 'white';
-            ctx.setLineDash([5, 15]);
-            ctx.beginPath();
-            ctx.moveTo(canvas.width / 2, 0);
-            ctx.lineTo(canvas.width / 2, canvas.height);
-            ctx.stroke();
-            ctx.setLineDash([]);
-
-            // Draw score
-            ctx.font = '48px Arial';
-            ctx.fillStyle = 'white';
-            ctx.textAlign = 'center';
-            ctx.fillText(gameState.score1.toString(), canvas.width / 4, 60);
-            ctx.fillText(gameState.score2.toString(), (canvas.width * 3) / 4, 60);
-
-            // Draw paddles
-            ctx.fillStyle = 'white';
-            
-            // Left paddle
-            ctx.fillRect(
-                scaleX(0),
-                scaleY(gameState.paddle1.y - PADDLE_HEIGHT/2),
-                scaleX(PADDLE_WIDTH),
-                scaleY(PADDLE_HEIGHT)
-            );
-
-            // Right paddle
-            ctx.fillRect(
-                scaleX(100 - PADDLE_WIDTH),
-                scaleY(gameState.paddle2.y - PADDLE_HEIGHT/2),
-                scaleX(PADDLE_WIDTH),
-                scaleY(PADDLE_HEIGHT)
-            );
-
-            // Draw ball
-            ctx.beginPath();
-            ctx.arc(
-                scaleX(gameState.ball.x),
-                scaleY(gameState.ball.y),
-                scaleX(BALL_SEMIDIAMETER),
-                0,
-                Math.PI * 2
-            );
-            ctx.fillStyle = 'red';
-            ctx.fill();
-
-            // Draw game status
-            ctx.font = '24px Arial';
-            ctx.fillText(`Status: ${gameState.status}`, canvas.width / 2, 30);
-        }
-
-        // Handle paddle movement
-        document.addEventListener('keydown', (event) => {
-            if (!socket || socket.readyState !== WebSocket.OPEN) return;
-
-            let direction = 0;
-            if (event.key === 'ArrowUp') {
-                direction = -1;
-            } else if (event.key === 'ArrowDown') {
-                direction = 1;
-            }
-
-            if (direction !== 0) {
-                socket.send(JSON.stringify({
-                    type: 'movePaddle',
-                    direction: direction
-                }));
-            }
-        });
-
-        // Initial blank screen
-        ctx.fillStyle = '#1a1a1a';
-        ctx.fillRect(0, 0, canvas.width, canvas.height);
-        ctx.font = '24px Arial';
-        ctx.fillStyle = 'white';
-        ctx.textAlign = 'center';
-        ctx.fillText('Connect as Player 1 or Player 2 to start', canvas.width / 2, canvas.height / 2);
-    </script>
-</body>
-</html>
