+++ services/game_service/tsconfig.json
@@ -8,6 +8,6 @@
     "strict": true,
     "skipLibCheck": true,
     "outDir": "./dist",
-    "rootDir": "./"
+    "rootDir": "./src"
   }
 }
