--- services/game_service/types/messages.ts
@@ -1,5 +1,0 @@
-// types/messages.ts
-export interface PaddleMoveMessage {
-    type: 'movePaddle';
-    direction: 1 | -1;
-}
